+++ cli/commands/metadata_diff.go
@@ -5,9 +5,11 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/hasura/graphql-engine/cli/migrate"
+	"github.com/sirupsen/logrus"
 
 	"github.com/aryann/difflib"
 	"github.com/hasura/graphql-engine/cli"
@@ -16,6 +18,11 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
+)
+
+const (
+	folderType string = "folder"
+	fileType   string = "file"
 )
 
 type MetadataDiffOptions struct {
@@ -47,7 +54,10 @@ By default, shows changes between exported metadata file and server metadata.`,
   hasura metadata diff local_metadata.yaml
 
   # Show changes between metadata from metadata.yaml and metadata_old.yaml:
-  hasura metadata diff metadata.yaml metadata_old.yaml
+  hasura metadata diff metadata.yaml ../old/metadata_old.yaml
+	
+  # Show changes among multiple metadata folders (v2 config only):
+  hasura metadata diff metadata1 metadata2
 
   # Apply admin secret for Hasura GraphQL Engine:
   hasura metadata diff --admin-secret "<admin-secret>"
@@ -65,7 +75,7 @@ By default, shows changes between exported metadata file and server metadata.`,
 }
 
 func (o *MetadataDiffOptions) runv2(args []string) error {
-	messageFormat := "Showing diff between %s and %s..."
+	messageFormat := "Showing diff between %s directory and %s..."
 	message := ""
 
 	switch len(args) {
@@ -74,24 +84,28 @@ func (o *MetadataDiffOptions) runv2(args []string) error {
 		message = fmt.Sprintf(messageFormat, o.Metadata[0], "the server")
 	case 1:
 		// 1 arg, diff given directory and the metadata on server
-		err := checkDir(args[0])
+		err := checkDirAndFile(args[0], folderType)
 		if err != nil {
 			return err
 		}
 		o.Metadata[0] = args[0]
 		message = fmt.Sprintf(messageFormat, o.Metadata[0], "the server")
 	case 2:
-		err := checkDir(args[0])
+		err := checkDirAndFile(args[0], folderType)
 		if err != nil {
 			return err
 		}
 		o.Metadata[0] = args[0]
-		err = checkDir(args[1])
+		err = checkDirAndFile(args[1], folderType)
 		if err != nil {
 			return err
 		}
 		o.Metadata[1] = args[1]
-		message = fmt.Sprintf(messageFormat, o.Metadata[0], o.Metadata[1])
+		if args[0] == args[1] {
+			return errors.New("directories passed are the same")
+		}
+		messageEnd := fmt.Sprintf("the %s directory", o.Metadata[1])
+		message = fmt.Sprintf(messageFormat, o.Metadata[0], messageEnd)
 	}
 	o.EC.Logger.Info(message)
 	var oldYaml, newYaml []byte
@@ -139,7 +153,7 @@ func (o *MetadataDiffOptions) runv2(args []string) error {
 		return errors.Wrap(err, "cannot unmarshal local metadata")
 	}
 
-	printDiff(string(oldYaml), string(newYaml), o.Output)
+	printDiff(string(oldYaml), string(newYaml), o.Output, o.EC.Logger)
 	return nil
 }
 
@@ -159,13 +173,41 @@ func (o *MetadataDiffOptions) runv1(args []string) error {
 		message = fmt.Sprintf(messageFormat, filename, "the server")
 	case 1:
 		// 1 arg, diff given filename and the metadata on server
-		o.Metadata[0] = args[0]
-		message = fmt.Sprintf(messageFormat, args[0], "the server")
+		// we are modifying this since we got
+		modifiedFilePath := includeMetadataFileInPath(args[0])
+		if modifiedFilePath == "" {
+			err := checkDirAndFile(args[0], fileType)
+			if err != nil {
+				return err
+			}
+			o.Metadata[0] = args[0]
+		} else {
+			o.Metadata[0] = modifiedFilePath
+		}
+		message = fmt.Sprintf(messageFormat, o.Metadata[0], "the server")
 	case 2:
 		// 2 args, diff given filenames
-		o.Metadata[0] = args[0]
-		o.Metadata[1] = args[1]
-		message = fmt.Sprintf(messageFormat, args[0], args[1])
+		modifiedFilePath := includeMetadataFileInPath(args[0])
+		if modifiedFilePath == "" {
+			err := checkDirAndFile(args[0], fileType)
+			if err != nil {
+				return err
+			}
+			o.Metadata[0] = args[0]
+		} else {
+			o.Metadata[0] = modifiedFilePath
+		}
+		modifiedFilePath = includeMetadataFileInPath(args[1])
+		if modifiedFilePath == "" {
+			err := checkDirAndFile(args[1], fileType)
+			if err != nil {
+				return err
+			}
+			o.Metadata[1] = args[1]
+		} else {
+			o.Metadata[1] = modifiedFilePath
+		}
+		message = fmt.Sprintf(messageFormat, o.Metadata[0], o.Metadata[1])
 	}
 
 	o.EC.Logger.Info(message)
@@ -198,7 +240,7 @@ func (o *MetadataDiffOptions) runv1(args []string) error {
 		return errors.Wrap(err, "cannot read file")
 	}
 
-	printDiff(string(oldYaml), string(newYaml), o.Output)
+	printDiff(string(oldYaml), string(newYaml), o.Output, o.EC.Logger)
 	return nil
 }
 
@@ -209,30 +251,83 @@ func (o *MetadataDiffOptions) Run() error {
 	return o.runv1(o.Args)
 }
 
-func printDiff(before, after string, to io.Writer) {
+// Move to using myers diffing algo
+func printDiff(before, after string, to io.Writer, logger *logrus.Logger) {
 	diffs := difflib.Diff(strings.Split(before, "\n"), strings.Split(after, "\n"))
 
-	for _, diff := range diffs {
-		text := diff.Payload
+	isThereADiff := false
 
+	for _, diff := range diffs {
 		switch diff.Delta {
-		case difflib.RightOnly:
-			fmt.Fprintf(to, "%s\n", ansi.Color(text, "green"))
 		case difflib.LeftOnly:
-			fmt.Fprintf(to, "%s\n", ansi.Color(text, "red"))
-		case difflib.Common:
-			fmt.Fprintf(to, "%s\n", text)
+		case difflib.RightOnly:
+			isThereADiff = true
+			break
 		}
 	}
+
+	if isThereADiff {
+		logger.Info("displaying diff...")
+		fmt.Fprintf(to, "\n")
+		for _, diff := range diffs {
+			text := diff.Payload
+			// TODO: Probably printing just these is not enough,
+			// we may need to print 2-3 line before and after these
+			// changes to provide more context to the users
+			switch diff.Delta {
+			case difflib.RightOnly:
+				fmt.Fprintf(to, "%s\n", ansi.Color(text, "green"))
+			case difflib.LeftOnly:
+				fmt.Fprintf(to, "%s\n", ansi.Color(text, "red"))
+			}
+		}
+		return
+	}
+
+	logger.Info("no diff")
 }
 
-func checkDir(path string) error {
+// checkDirAndFile supports either "file" or "folder"
+func checkDirAndFile(path, pathType string) error {
 	file, err := os.Stat(path)
+
 	if err != nil {
+		if os.IsNotExist(err) {
+			return fmt.Errorf("this %s does not exist", pathType)
+		}
 		return err
 	}
-	if !file.IsDir() {
-		return fmt.Errorf("metadata diff only works with folder but got file %s", path)
+
+	if !file.IsDir() && pathType == folderType {
+		// This is meant for v2 config
+		return fmt.Errorf("metadata diff only works with metadata directories but got %s", path)
 	}
+
+	if file.IsDir() && pathType == fileType {
+		// This is meant for v1 config and no metadata.yml file wasn't found in the given dir.
+		return fmt.Errorf("metadata diff only works with metadata files but got %s", path)
+	}
+
 	return nil
+}
+
+func includeMetadataFileInPath(path string) string {
+	folder, err := os.Stat(path)
+	if err != nil {
+		return ""
+	}
+	if !folder.IsDir() {
+		return ""
+	}
+	files, err := ioutil.ReadDir(path)
+	if err != nil {
+		return ""
+	}
+	for _, file := range files {
+		currentFile := file.Name()
+		if currentFile == "metadata.yaml" || currentFile == "metadata.yml" {
+			return filepath.Join(path, currentFile)
+		}
+	}
+	return ""
 }